requirements.txt
discord.py
typing